requirements.txt
Flask==2.3.3
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
google-auth==2.23.4
Werkzeug==2.3.7
openai-whisper
googletrans==4.0.0rc1
gtts
firebase-admin
torch
torchvision
torchaudio
ffmpeg-python